W229: template transfer round six — +1 MATCH (6571), nd -1339, four new idioms
Sixth main-model template-transfer wave. Eight agents, 118 targets at nd 20-500 with the measured
floors excluded. 0 MATCHes lost, 0 new over-window, 0 regressions kept.

MATCH: `comuSasoi 003c2c50` nd51 -> 0 — the matched sibling's goto-shaped loop completion with a
`u32` completion boolean, which also turned the residual `daddiu` into retail's `addiu`.

**Four transferable idioms came out of this wave, all measured:**

**Mixed-ABI prototype parameter order.** When a function takes both integer and float arguments the
ORDER they are DECLARED in changes codegen even with the physical ABI registers unchanged. Our
prototypes carry the decompiler's guess. Reordering `fclDrawSprite 0040e930` to matched sibling
`FUN_0040e710`'s interleaving gave nd46 -> 14; `0040eef0` went nd198 -> 190 the same way and then
**-> 7** with one addressing-mode correction on top. That pairing — a prototype reorder unlocking the
function, then one absolute-alias fix finishing it — is the shape to aim for.

**The linked-list traversal shape**, and the clearest cluster of the campaign. Initialize the node
before the counter and bound, stage the bound once, and advance the node BEFORE incrementing the
counter. Found at `mt_evtMain 00367a20` (nd32 -> 8), then applied immediately to three neighbours:
`00367230` 173 -> 149, `003698a0` 224 -> 194, `003680e0` 292 -> 268.

**Explicit 16-bit narrowing.** Where retail emits an `andi ...,0xffff` we lack, the original source
masked explicitly: `arr[idx & 0xffff]` gave `btlFormation 002c3ce0` nd106 -> 17, and `param_5 &=
0xffff` with u16 typed indexing gave `002c4c80` nd280 -> 178. A maintained `u16` counter gave the
persona-table pair `003d7ac0`/`003d7da0` nd399 -> 341 each. **It is strictly per-site** — Y229Mdl
measured it at `mdlFile 00318d10` and it regressed 268 -> 296, and Y229Menu found retail
sign-extending the counter it was offered for, so both correctly declined.

**Default-FIRST cascade, again**, now in a third file: `k_footstep 001de630` nd240 -> 195 and
`001ded40` nd355 -> 309 by putting the status body before the default body.

**A pragma that was helping in wave 161 is now actively hurting.** `btlFormation 002bc9c0` carried
`#pragma opt_loop_invariants on` annotated "removing this worsens nd562 -> nd618 - measured W161".
That annotation is stale: fifty waves of source repair later, deleting the pragma gives **nd484 ->
218** with zero MATCHes lost anywhere in the file. A tree-wide probe (`build/wip/W230_pragmaprobe.py`,
census in `build/wip/W230_probe.txt`) now measures all 637 first-party optimization-pragma directives
this way, and **19 first-party functions are in the same position.** It also shows what the rest of
them really are: 265 MATCHes sit in identical-nd-delta clusters — btlFormation alone has 13 functions
that all lose exactly 44 points, 9 that lose 158, and 6 that lose 43 — which means a handful of
shared source defects are being compensated dozens of times over. Those clusters are the wave 230
target, and each one closed removes both a defect and an honesty violation.

Also kept: `mt_evtMenu 0036fa80` nd279 -> 110 (the file's `_abs` alias spelling reproducing retail's
absolute LUI/LO formation), `mt_evtCustomEvent 0039c0f0` nd495 -> 434, `bpc_main 00244770` nd32 -> 4,
`mt_evtCustomEvent 003be020` nd219 -> 199 (absolute alias, then `u16` -> `s16` for signed `lh`),
`comuTimerSequence 003c2ee0` nd131 -> 130 (`u32` -> `s32` giving retail's signed `slti`).

**Y229Data returned twelve rigorous negatives and zero keeps: the datCalc/datCalendar band is mined
out at this depth.** It declined several targets by reading the measurement history already recorded
in the source rather than repeating it, which is exactly right. Y229Mdl confirmed mdlFile's low-diff
band is entirely call-argument setup order, and correctly refused to reorder a `primitive.h` public
prototype because the callers live in three files it did not own — that reorder is orchestrator work.
Repaired `fclDrawSprite.c`, which an agent had converted from LF to CRLF wholesale.

+++ src/Event/Fcl/fclCombineMisc.c
@@ -2058,7 +2058,7 @@ u32 FUN_003d7ac0(u8 param_1,u8 param_2,u16 *param_3)
 
   u32 uVar9;
 
-  u32 uVar10;
+  u16 uVar10;
 
   int iVar11;
 
@@ -2238,7 +2238,7 @@ u32 FUN_003d7da0(u8 param_1,u8 param_2,u16 *param_3)
 
   u32 uVar9;
 
-  u32 uVar10;
+  u16 uVar10;
 
   int iVar11;
